atlas_video-processing/ingest/drone_client/requirements.txt
# Add any specific Python dependencies here
# For now, we'll keep it minimal.
# Later, this might include libraries for video capture, networking, etc.
# For example:
# opencv-python
# paho-mqtt
kafka-python